chore: refactor NavigationPane to its own file
