+++ src/Frontend/Panes/Lobbies/ConfigurationPane.tsx
@@ -2,30 +2,9 @@ import { EthAddress } from '@darkforest_eth/types';
 import _ from 'lodash';
 import React, { useEffect, useReducer, useState } from 'react';
 import { Route, Switch, useRouteMatch } from 'react-router-dom';
-import { Btn } from '../../Components/Btn';
-import { Spacer, Title } from '../../Components/CoreUI';
-import { MythicLabelText } from '../../Components/Labels/MythicLabel';
-import { LoadingSpinner } from '../../Components/LoadingSpinner';
 import { Modal } from '../../Components/Modal';
-import { Row } from '../../Components/Row';
-import { TextPreview } from '../../Components/TextPreview';
-import { AdminPermissionsPane } from './AdminPermissionsPane';
-import { ArtifactSettingsPane } from './ArtifactSettingsPane';
-import { CaptureZonesPane } from './CaptureZonesPane';
-import { GameSettingsPane } from './GameSettingsPane';
-import {
-  ButtonRow,
-  ConfigDownload,
-  ConfigUpload,
-  LinkButton,
-  LobbiesPaneProps,
-  NavigationTitle,
-  Warning,
-} from './LobbiesUtils';
+import { ConfigDownload, ConfigUpload, Warning } from './LobbiesUtils';
 import { MinimapConfig } from './MinimapUtils';
-import { PlanetPane } from './PlanetPane';
-import { CreatePlanetPane } from './CreatePlanetPane';
-import { PlayerSpawnPane } from './PlayerSpawnPane';
 import {
   InvalidConfigError,
   LobbyConfigAction,
@@ -35,198 +14,9 @@ import {
   LobbyInitializers,
   toInitializers,
 } from './Reducer';
-import { SnarkPane } from './SnarkPane';
-import { SpaceJunkPane } from './SpaceJunkPane';
-import { SpaceTypeBiomePane } from './SpaceTypeBiomePane';
-import { WorldSizePane } from './WorldSizePane';
-import { TargetPlanetPane } from './TargetPlanetPane';
 import { PlanetCreator } from '../../../Backend/Utils/PlanetCreator';
-
-interface PaneConfig {
-  title: string;
-  shortcut: string;
-  path: string;
-  Pane: (props: LobbiesPaneProps) => JSX.Element;
-}
-
-const panes: ReadonlyArray<PaneConfig> = [
-  {
-    title: 'Game settings',
-    shortcut: `1`,
-    path: '/settings/game',
-    Pane: (props: LobbiesPaneProps) => <GameSettingsPane {...props} />,
-  },
-  {
-    title: 'World size',
-    shortcut: `2`,
-    path: '/settings/world',
-    Pane: (props: LobbiesPaneProps) => <WorldSizePane {...props} />,
-  },
-  {
-    title: 'Space type & Biome',
-    shortcut: `3`,
-    path: '/settings/space',
-    Pane: (props: LobbiesPaneProps) => <SpaceTypeBiomePane {...props} />,
-  },
-  {
-    title: 'Planets',
-    shortcut: `4`,
-    path: '/settings/planet',
-    Pane: (props: LobbiesPaneProps) => <PlanetPane {...props} />,
-  },
-  {
-    title: 'Player spawn',
-    shortcut: `5`,
-    path: '/settings/spawn',
-    Pane: (props: LobbiesPaneProps) => <PlayerSpawnPane {...props} />,
-  },
-  {
-    title: 'Space junk',
-    shortcut: `6`,
-    path: '/settings/junk',
-    Pane: (props: LobbiesPaneProps) => <SpaceJunkPane {...props} />,
-  },
-  {
-    title: 'Capture zones',
-    shortcut: `7`,
-    path: '/settings/zones',
-    Pane: (props: LobbiesPaneProps) => <CaptureZonesPane {...props} />,
-  },
-  {
-    title: 'Artifacts',
-    shortcut: `8`,
-    path: '/settings/artifact',
-    Pane: (props: LobbiesPaneProps) => <ArtifactSettingsPane {...props} />,
-  },
-  {
-    title: 'Admin permissions',
-    shortcut: `9`,
-    path: '/settings/admin',
-    Pane: (props: LobbiesPaneProps) => <AdminPermissionsPane {...props} />,
-  },
-  {
-    title: 'Advanced: Snarks',
-    shortcut: `0`,
-    path: '/settings/snark',
-    Pane: (props: LobbiesPaneProps) => <SnarkPane {...props} />,
-  },
-  {
-    title: 'Target planets',
-    shortcut: `-`,
-    path: '/settings/arena',
-    Pane: (props: LobbiesPaneProps) => <TargetPlanetPane {...props} />,
-  },
-  {
-    title: 'Admin planets',
-    shortcut: `+`,
-    path: '/settings/create',
-    Pane: (props: LobbiesPaneProps) => <CreatePlanetPane {...props} />,
-  },
-] as const;
-
+import { ConfigurationNavigation } from './WorldSettingsNavPane';
 type Status = 'creating' | 'created' | 'errored' | undefined;
-
-function ConfigurationNavigation({
-  error,
-  lobbyAddress,
-  status,
-  statusMessage,
-  onCreate,
-  createPlanets,
-  config,
-}: {
-  error: string | undefined;
-  lobbyAddress: EthAddress | undefined;
-  status: Status;
-  statusMessage: string;
-  onCreate: () => Promise<void>;
-  createPlanets: () => Promise<void>;
-  config: LobbyConfigState;
-}) {
-  const buttons = _.chunk(panes, 2).map(([fst, snd], idx) => {
-    return (
-      // Index key is fine here because the array is stable
-      <ButtonRow key={idx}>
-        {fst && (
-          <LinkButton to={fst.path} shortcut={fst.shortcut}>
-            {fst.title}
-          </LinkButton>
-        )}
-        {snd && (
-          <LinkButton to={snd.path} shortcut={snd.shortcut}>
-            {snd.title}
-          </LinkButton>
-        )}
-      </ButtonRow>
-    );
-  });
-
-  const url = `${window.location.origin}/play/${lobbyAddress}`;
-
-  let lobbyContent;
-  if (status === 'created' && lobbyAddress) {
-    lobbyContent = (
-      <>
-        {config.ADMIN_PLANETS.currentValue.length > 0 && (
-          <>
-            <Btn size='stretch' onClick={createPlanets}>
-              Create Planets
-            </Btn>
-            <Row />
-          </>
-        )}
-        <Btn size='stretch' onClick={() => window.open(url)}>
-          Launch Lobby
-        </Btn>
-        <Row>
-          {/* Stealing MythicLabelText because it accepts variable text input */}
-          <MythicLabelText style={{ margin: 'auto' }} text='Your lobby is ready!' />
-        </Row>
-        <Row>
-          <span style={{ margin: 'auto' }}>
-            You can also share the direct url with your friends:
-          </span>
-        </Row>
-        {/* Didn't like the TextPreview jumping, so I'm setting the height */}
-        <Row style={{ height: '30px' } as CSSStyleDeclaration & React.CSSProperties}>
-          <TextPreview
-            style={{ margin: 'auto' }}
-            text={url}
-            unFocusedWidth='50%'
-            focusedWidth='100%'
-          />
-        </Row>
-      </>
-    );
-  }
-
-  const createDisabled = status === 'creating' || status === 'created';
-  const creating = status === 'creating' || (status === 'created' && !lobbyAddress);
-  const created = (status === 'created' && lobbyAddress);
-  return (
-    <>
-      <Title slot='title'>Customize Lobby</Title>
-      <div>
-        Welcome Cadet! You can launch a copy of Dark Forest from this UI. We call this a Lobby.
-        <Spacer height={12} />
-        All settings will be defaulted to the same configuration of the main contract you are
-        copying. However, you can change any of those settings through the buttons below!
-        <Spacer height={12} />
-      </div>
-      {buttons}
-      <Spacer height={20} />
-      {!created && (
-        <Btn size='stretch' onClick={onCreate}>
-          {creating ? <LoadingSpinner initialText={statusMessage} /> : 'Create Lobby'}
-        </Btn>
-      )}
-      <Row>
-        <Warning>{error}</Warning>
-      </Row>
-      {lobbyContent}
-    </>
-  );
-}
 
 export function ConfigurationPane({
   modalIndex,
@@ -283,16 +73,6 @@ export function ConfigurationPane({
     config.PERLIN_THRESHOLD_3.currentValue,
     config.ADMIN_PLANETS.currentValue,
   ]);
-
-  const routes = panes.map(({ title, path, Pane }, idx) => {
-    return (
-      // Index key is fine here because the array is stable
-      <Route key={idx} path={`${root}${path}`}>
-        <NavigationTitle>{title}</NavigationTitle>
-        <Pane config={config} onUpdate={onUpdate} />
-      </Route>
-    );
-  });
 
   async function validateAndCreateLobby() {
     try {
@@ -353,7 +133,7 @@ export function ConfigurationPane({
   return (
     <Modal width='500px' initialX={100} initialY={100} index={modalIndex}>
       <Switch>
-        <Route path={root} exact={true}>
+        <Route path={`${root}`}>
           <ConfigurationNavigation
             error={error}
             lobbyAddress={lobbyAddress}
@@ -362,9 +142,9 @@ export function ConfigurationPane({
             onCreate={validateAndCreateLobby}
             createPlanets={createAndRevealPlanets}
             config={config}
+            onUpdate={onUpdate}
           />
         </Route>
-        {routes}
       </Switch>
       {/* Button this in the title slot but at the end moves it to the end of the title bar */}
       <ConfigDownload onError={setIoErr} address={lobbyAddress} config={config} />
